Add files via upload
Annales du patrimoine/Hawliyyat al-Turath (ADP) (Annals of Heritage) is a trilingual academic journal published with the participation of researchers in literature, languages and humanities. The journal is devoted to different areas of Algerian and Islamic cultural heritage: literature, language, spirituality, philosophy, art. It appears in Arabic, French and English once a year and distributed exclusively online.